+++ CHANGES.rst
@@ -39,6 +39,8 @@ parameter if the `next_masker` function is overriden.
 defined as class parameters, rather than being hard-coded.
 - Add optional `ax` parameter to py:func:`~pambox.speech.experiment.plot_results`.
 - Function py:func:`~pambox.speech.experiment.pred_to_pc` can now convert prediction to intelligibility for a specific model.
+- Possibility to force the audio file to be mono when loading file in
+py:class:`pambox.speech.material.Material`.
 
 Performance
 -----------